+++ libraries/resource_providers/api200/ethernet_network_provider.rb
@@ -41,7 +41,7 @@ module OneviewCookbook
       end
 
       def reset_connection_template
-        @item.retrieve!
+        @item.retrieve! || raise("#{@resource_name} '#{@name}' not found!")
         klass = resource_named(:ConnectionTemplate)
         update_connection_template(bandwidth: klass.get_default(@item.client)['bandwidth'])
       end