+++ spec/unit/resources/enclosure_group/set_script_spec.rb
@@ -17,4 +17,9 @@ describe 'oneview_test::enclosure_group_set_script' do
     expect_any_instance_of(OneviewSDK::EnclosureGroup).to receive(:set_script).with('hello, world!').and_return(true)
     expect(real_chef_run).to set_oneview_enclosure_group_script('EnclosureGroup3')
   end
+
+  it 'fails if the resource is not found' do
+    expect_any_instance_of(OneviewSDK::EnclosureGroup).to receive(:retrieve!).and_return(false)
+    expect { real_chef_run }.to raise_error(RuntimeError, /not found/)
+  end
 end