+++ spec/unit/resources/rack/remove_from_rack_spec.rb
@@ -47,4 +47,9 @@ describe 'oneview_test::rack_remove_from_rack' do
     expect_any_instance_of(OneviewSDK::Rack).to_not receive(:update)
     expect(real_chef_run).to remove_oneview_rack_from_rack('Rack1')
   end
+
+  it 'fails if the resource is not found' do
+    expect_any_instance_of(OneviewSDK::Rack).to receive(:retrieve!).and_return(false)
+    expect { real_chef_run }.to raise_error(RuntimeError, /not found/)
+  end
 end