+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Adds support to API300 HPE Synergy Image Streamer resources:
 
 Bug Fixes:
 - [#93](https://github.com/HewlettPackard/oneview-chef/issues/93) oneview_storage_system should not try to update the name
- 
+- [#220](https://github.com/HewlettPackard/oneview-chef/issues/220) Raise "not found" error after failed retrieval of resource
+
 Enhancements:
 - [#218](https://github.com/HewlettPackard/oneview-chef/issues/218) connection_template resource does not support NetworkSet, FCoENetwork and FCNetwork